Former-commit-id: 90ff95df130c350fea978cab7d74e3e16d83a9cd

+++ root/deploy.sh
@@ -1,17 +1,15 @@
 #!/bin/bash
 # This script will attempt to symlink the files in these folders to the root directory, using stow.
 # conflicts by default cause the script to stop
-# if an argument exists, will force-resolve conflicts, storing old files in a folder ../BASEconflictFiles
+# if an argument exists, will force-resolve conflicts, storing old files in a folder ../ROOTconflictFiles
 
 if [ "$(id -u)" != "0" ]; then
-   echo "This script must be run as root" 1>&2
+   echo "This script must be run as root." 1>&2
    exit 1
 fi
 
 #clear old conflicts file if it exists
-if [ -f conflicts.txt ]; then
-    rm conflicts.txt
-fi
+[ -f conflicts.txt ] && rm conflicts.txt
 
 for i in $(ls -d */); do
     #see if there will be any file conflicts:
@@ -19,17 +17,15 @@ for i in $(ls -d */); do
 done
 
 #remove conflicts if it is empty(file was created because we directed output to it(even if that output is empty):
-if [ ! -s conflicts.txt ]; then
-    rm conflicts.txt
-fi
+if [ ! -s conflicts.txt ] && rm conflicts.txt
 
 if [ -f conflicts.txt ];then
     if [ -z "$1" ]; then
-        echo "STOPPING DUE TO FILE CONFLICTS, check conflicts.txt"
+        echo "stopping due to file conficts, check conflicts.txt"
         exit 0
     fi
 
-    mkdir ../BASEconflictFiles 2> /dev/null
+    mkdir ../ROOTconflictFiles 2> /dev/null
 
     grep -oE "directory:.*" conflicts.txt | while read -r conflictName; do
         #cut off the 'directory:' text
@@ -37,12 +33,12 @@ if [ -f conflicts.txt ];then
         #make a dir if needed
         cfDir=$(echo $conflictFile | grep -oE ".+/")
         if [ ! -z "$cfDir" ]; then
-            mkdir -p "../BASEconflictFiles/$cfDir"
+            mkdir -p "../ROOTconflictFiles/$cfDir"
         fi
-        mv /$conflictFile ../BASEconflictFiles/$conflictFile
+        mv /$conflictFile ../ROOTconflictFiles/$conflictFile
     done
 
-    echo conflicts moved to ../BASEconflictFiles/$conflictFile
+    echo conflicts moved to ../ROOTconflictFiles/$conflictFile
 fi
 
 rm conflicts.txt 2> /dev/null
